fixed typo and renamed storage class instance folder
