+++ lib/entities/Storage/index.js
@@ -1,4 +1,4 @@
-export {default as Storage} from './singleton';
+export {default as Storage} from './instance';
 export {default as DEFAULT_STORAGE_CONFIG} from './constants/defaultStorageConfig';
 export {default as saveNotification} from './utils/saveNotification';
 export {default as removeOldestNotification} from './utils/removeOldestNotification';
