+++ lib/entities/Storage/instance/index.js
@@ -1,0 +1,7 @@
+import MMKVStorage from '@janiscommerce/app-storage';
+
+const Storage = new MMKVStorage({
+  id: 'push-notification',
+});
+
+export default Storage;
