--- lib/entities/Storage/singleton/index.js
@@ -1,7 +1,0 @@
-import MMMKVStorage from '@janiscommerce/app-storage';
-
-const Storage = new MMMKVStorage({
-  id: 'push-notification',
-});
-
-export default Storage;
