+++ lib/entities/Storage/utils/getStoredNotification/index.js
@@ -1,4 +1,4 @@
-import Storage from '../../singleton';
+import Storage from '../../instance';
 
 /**
  * Retrieves a stored notification by type and messageId.
