+++ lib/entities/Storage/utils/hasNotificationStore/index.js
@@ -1,4 +1,4 @@
-import Storage from '../../singleton';
+import Storage from '../../instance';
 
 /**
  * Checks whether there is an existing store entry for the given notification type.
