+++ lib/entities/Storage/utils/removeOldestNotification/index.js
@@ -1,4 +1,4 @@
-import Storage from '../../singleton';
+import Storage from '../../instance';
 
 /**
  * Removes the oldest (first) stored notification for a given event/type.
