+++ lib/entities/Storage/utils/removeStoredNotification/index.js
@@ -1,4 +1,4 @@
-import Storage from '../../singleton';
+import Storage from '../../instance';
 
 /**
  * Removes a stored notification matching the given type and messageId.
