+++ lib/entities/Storage/utils/saveNotification/index.js
@@ -1,4 +1,4 @@
-import Storage from '../../singleton';
+import Storage from '../../instance';
 import {isObject} from '../../../../utils';
 
 /**
